+++ src/NosCore.MasterServer/Controllers/FriendController.cs
@@ -20,7 +20,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ChickenAPI.Packets.Enumerations;
+using System.Threading.Tasks;
+using NosCore.Packets.Enumerations;
 using Microsoft.AspNetCore.Mvc;
 using NosCore.Core;
 using NosCore.Core.HttpClients.ConnectedAccountHttpClient;
@@ -57,13 +58,13 @@ namespace NosCore.MasterServer.Controllers
 
 
         [HttpPost]
-        public LanguageKey AddFriend([FromBody] FriendShipRequest friendPacket)
+        public async Task<LanguageKey> AddFriend([FromBody] FriendShipRequest friendPacket)
         {
-            var character = _connectedAccountHttpClient.GetCharacter(friendPacket.CharacterId, null);
-            var targetCharacter = _connectedAccountHttpClient.GetCharacter(friendPacket.FinsPacket.CharacterId, null);
+            var character = await _connectedAccountHttpClient.GetCharacter(friendPacket.CharacterId, null);
+            var targetCharacter = await _connectedAccountHttpClient.GetCharacter(friendPacket.FinsPacket.CharacterId, null);
             var friendRequest = _friendRequestHolder.FriendRequestCharacters.Where(s =>
-                (s.Value.Item2 == character.Item2.ConnectedCharacter.Id) &&
-                (s.Value.Item1 == targetCharacter.Item2.ConnectedCharacter.Id)).ToList();
+                (s.Value.Item2 == character.Item2?.ConnectedCharacter.Id) &&
+                (s.Value.Item1 == targetCharacter.Item2?.ConnectedCharacter.Id)).ToList();
             if ((character.Item2 != null) && (targetCharacter.Item2 != null))
             {
                 if (character.Item2.ChannelId != targetCharacter.Item2.ChannelId)
@@ -141,7 +142,7 @@ namespace NosCore.MasterServer.Controllers
         }
 
         [HttpGet]
-        public List<CharacterRelationStatus> GetFriends(long id)
+        public async Task<List<CharacterRelationStatus>> GetFriends(long id)
         {
             var charList = new List<CharacterRelationStatus>();
             var list = _characterRelationDao
@@ -150,9 +151,9 @@ namespace NosCore.MasterServer.Controllers
             {
                 charList.Add(new CharacterRelationStatus
                 {
-                    CharacterName = _characterDao.FirstOrDefault(s => s.CharacterId == rel.RelatedCharacterId).Name,
+                    CharacterName = _characterDao.FirstOrDefault(s => s.CharacterId == rel.RelatedCharacterId)?.Name,
                     CharacterId = rel.RelatedCharacterId,
-                    IsConnected = _connectedAccountHttpClient.GetCharacter(rel.RelatedCharacterId, null).Item1 != null,
+                    IsConnected = (await _connectedAccountHttpClient.GetCharacter(rel.RelatedCharacterId, null)).Item1 != null,
                     RelationType = rel.RelationType,
                     CharacterRelationId = rel.CharacterRelationId
                 });
@@ -166,9 +167,17 @@ namespace NosCore.MasterServer.Controllers
         {
             var rel = _characterRelationDao.FirstOrDefault(s =>
                 (s.CharacterRelationId == id) && (s.RelationType == CharacterRelationType.Friend));
+            if (rel == null)
+            {
+                return NotFound();
+            }
             var rel2 = _characterRelationDao.FirstOrDefault(s =>
                 (s.CharacterId == rel.RelatedCharacterId) && (s.RelatedCharacterId == rel.CharacterId) &&
                 (s.RelationType == CharacterRelationType.Friend));
+            if (rel2 == null)
+            {
+                return NotFound();
+            }
             _characterRelationDao.Delete(rel);
             _characterRelationDao.Delete(rel2);
             return Ok();
